refactor: #209 두 개가 존재했언 장소 dto를 하나로 통합하여 관리하도록 수정
